Added animations and UI beautification to the login flow. Fixed minor bugs in app
Changes this commit:
1) Added animations for fragment transactions in login flow
2) Added better empty field prompts for UX.
3) Fixed a minor bug where app crashes if user does not select a profile
picture on register page.

The app's UI flow has been spiffed up with beautiful animations and
reworked some of the UI elements present in the login flow XML Files.
Added in error prompt for editable fields for better user experience
instead of a toast appearing on the bottom of the screen.

If a user did not select the profile picture during register, the URI
imagePath that is held as an instance variable would be null and causes
a NPE when being passed into the firebase storage reference causing the
app to crash and the user being magically logged in after app restarts.
Fixed that with a null check on the profile picture and to upload a
random default photo of a banana if user does not select a picture.