Added more animations to transitions and fixed minor bugs.
Changes this commit:
1) Added more animations and loading transitions between screens
2) Added resources for loading screens and images
3) Fixed bugs where API call returning empty results results in crash

Added way more animations to the whole app both in the login flow and
the actual app itself so transitions are not so jarring. Also changed
pictures for some of the items in recycler view. Displays a loading
icon now instead of a stupid banana. If no results are found, a
corresponding page is also shown.

Now the app will call the API to fetch results if there are no results
in the database currently existing and add them to the database. If
even the API results in a blank page, it used to crash the app with
NullPointerException. But that has been cleared and it simply displays
the no results found page now.