Completed feature of sorting by most visits on results page.
Changes made this commit:
1) Sorting by most popular(visits) on results page
2) Added extra methods to Ranker for above mentioned ranking
3) Added menu item in results page to enable users to sort
4) Added loading progress bar to web view

The last feature of sorting by most visits to a link has been done.
Users can use the overflow menu to find the sorting button to
sort the results page by.

The webview is now able to display a progress bar to inform users
of when the loading will be done. Also, a web settings change has been
done to enable DOM storage on the phone. This was to prevent
null exceptions being thrown by the chromium web browser when it tries
to access the DOM storage for persistent data.