Fixed bug of favourites fragment not appearing.
Changed made this commit:
1) Fixed bug of favourites fragment getting null pointer

The issue with the fragment was that the completeLoadingScreen
method was being called before the views were mounted onto
the root view, which caused a null pointer exception onto
the group that was holding the loading view. The proper
time to call the mentioned method was when the adapter
was set up properly and the data has been fetched from
the database for the user's favourites. Thus, it has
been shifted over to the callback for the database
completion of query.