Fixed bugs present in the app.
Changed made this commit:
1) Fixed bug where searchview displaying nothing
2) Fixed bug where delete button crashes app
3) Fixed bug where delete button can be continually pressed
4) Added display of no favourites found page if no user has no fav.

When the adapter for the homefragment's recycler view was constructed,
it was passed in a list of ingredients. However, internally in the
IngredientAdapter class, the passed in list of ingredients was being
kept in two separate places, the main list used as the data set and a
full ingredient list that is being used for filtering. When this list
was initially passed in, it was empty due to database callback not
finishing. When the callback came back, the adapter was notified of
a data set changed, however, only the main list being used as the
data set was updated and informed. The full ingredient list was not
informed of this change and remains as an empty list. Thus, when the
filter was called, the full ingredient list that was used to aid in
filtering was empty and all subsequent results will thus be empty. The
fix was to simply shift the initialisation of the adapter over to when
the database callback has finished. This makes sense too since the
adapter will have nothing to display if the callback has not finished
and there would be no fear of adapter methods being called if database
has not finished its fetching.

Another minor bug was where the delete button causes crashes if u press
from a ConcurrentModificationError. This was due to the use of a
remove method inside a enhanced for loop(which internally uses an
iterator), which is a classic example of a CME exception being thrown.
Just replacing the enhanced for loop with a basic for loop will solve
such issues.

The last bug was where the delete button could be pressed even if no
ingredients was checked. This was caused by the checkedIngredients
list that was being stored inside the main recyclerview's adapter.
The list was never cleared when the removal of ingredients has
completed so it still held references to ingredients internally which
caused the isEmpty check to fail for the removeBtn callback method
resulting in multiple ingredients being removed. The solution was
simply to initialised a new list when removal was complete done so
using a notify method in the adapter class.