Fixed bugs present in the app.
Changes made this commit:
1)Fixed Results Fragment not displaying properly when the recipe in the query has already been favourited by user.
2)Fixed bug in Favourites Fragment where if user had multiple favourite recipes and wanted to delete them, it would leave a recipe still present in the page even after it was deleted.

For both bug fixes, calls to the database were originally made in the adapter classes. However, completion of these calls happened only after the viewholders were set up. As certain aspects of the views depended on the data from the database, this resulted in incorrect display of data in the viewholders.

The bug was fixed by changing the constructor to include the data that was needed as one of its parameters and instead making the call to the database in the Fragment classes instead.